backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for Deception Manager Game"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi>=0.115.0",
    "gotrue>=2.12.0",
    "loguru>=0.7.2",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.10",
    "redis>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "supabase>=2.11.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=14.0",
    "requests>=2.32.5",
    "cloudscraper>=1.2.71",
    "pillow",
    "psycopg2-binary>=2.9.11",
]
